requirements.txt
langchain==0.2.14
langchain-community==0.2.12
langchain-openai==0.2.3
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
pypdf==4.2.0
ollama==0.3.1
python-dotenv==1.0.1
fastapi==0.111.0
uvicorn[standard]==0.30.0